# 09_api_spec/openapi_v1.yaml
openapi: "3.0.3"
info:
  title: SCVRI Platform API
  description: |
    Supply Chain Visibility & Risk Intelligence Platform — Full REST API Specification
    All endpoints require Bearer JWT authentication.
    Tenant context is always derived from the JWT claim; never from request body.
    API versioning: URI-based (/api/v1/...). Breaking changes increment version to /api/v2/.
    Rate limits: returned in response headers X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset.
  version: "1.0.0"
  contact:
    email: api-support@scvri.io
  license:
    name: Proprietary
servers:
  - url: https://api.scvri.io/api/v1
    description: Production
  - url: https://api.staging.scvri.io/api/v1
    description: Staging
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Pagination:
      type: object
      properties:
        total: {type: integer}
        page: {type: integer}
        page_size: {type: integer}
        has_next: {type: boolean}
...
    AlertSummary:
      type: object
      properties:
        alert_id: {type: string, format: uuid}
        severity: {type: string, enum: [critical, high, medium, low]}
        entity_type: {type: string}
        entity_id: {type: string}
        title: {type: string}
        status: {type: string}
        generated_at: {type: string, format: date-time}
        acknowledged_at: {type: string, format: date-time, nullable: true}
        assigned_to: {type: string, format: uuid, nullable: true}
        delivery_latency_ms: {type: integer, nullable: true}
...
    User:
      type: object
      properties:
        user_id: {type: string, format: uuid}
        email: {type: string, format: email}
        display_name: {type: string}
        user_type: {type: string}
        status: {type: string}
        roles: {type: array, items: {type: string}}
        mfa_enrolled: {type: boolean}
        last_login_at: {type: string, format: date-time, nullable: true}
        created_at: {type: string, format: date-time}
paths:
...
  /alerts:
    get:
      tags: [Alerts]
      summary: Alert inbox (ALERT-FR-005)
      parameters:
        - in: query
          name: status
          schema: {type: array, items: {type: string}}
        - in: query
          name: severity
          schema: {type: array, items: {type: string}}
        - in: query
          name: page
          schema: {type: integer, default: 1}
        - in: query
          name: page_size
          schema: {type: integer, default: 50}
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Pagination'
                  - type: object
                    properties:
                      items: {type: array, items: {$ref: '#/components/schemas/AlertSummary'}}
                      unread_count: {type: integer}
  /alerts/{alertId}/acknowledge:
    post:
      tags: [Alerts]
      parameters:
        - in: path
          name: alertId
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSummary'
      x-audit: "alert.acknowledge"
  /alerts/{alertId}/resolve:
    post:
      tags: [Alerts]
      parameters:
        - in: path
          name: alertId
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [resolution_notes]
              properties:
                resolution_notes: {type: string, minLength: 10}
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlertSummary'
      x-audit: "alert.resolve"
...
  /admin/users:
    get:
      tags: [Administration]
      summary: List users (IAM-FR-001)
      x-required-role: "it_administrator"
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags: [Administration]
      summary: Invite user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, display_name, role_ids]
              properties:
                email: {type: string, format: email}
                display_name: {type: string}
                role_ids: {type: array, items: {type: string, format: uuid}}
                data_scope: {type: object}
      responses:
        "201":
          description: Invitation sent
      x-audit: "user.invite"
      x-required-role: "it_administrator"
  /admin/users/{userId}:
    put:
      tags: [Administration]
      parameters:
        - in: path
          name: userId
          required: true
          schema: {type: string, format: uuid}
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status: {type: string, enum: [active, locked, deactivated]}
                role_ids: {type: array, items: {type: string, format: uuid}}
                data_scope: {type: object}
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      x-audit: "user.update"
      x-required-role: "it_administrator"
...